Cleans up LoginScreen further; implements "Logging in..." message
